+++ modules/hanna-react/src/utils/browserSide.ts
@@ -1,0 +1,1 @@
+export { useIsBrowserSide, useIsServerSide } from '@hugsmidjan/react/hooks';